fix tests and lib/foldersummary
